move safesync package to interal dir
